engines/check/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
boto3>=1.34.0
botocore>=1.34.0
psycopg2-binary>=2.9.9
pyyaml>=6.0
python-dotenv>=0.21.0
httpx>=0.25.2
requests>=2.31.0
cryptography>=41.0.0
kubernetes>=28.1.0

# ── OpenTelemetry instrumentation ────────────────────────────────────────────
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0
opentelemetry-instrumentation-psycopg2>=0.46b0
opentelemetry-instrumentation-logging>=0.46b0
